Создает метод для получения отсортированных по времени событий в модели.
